Added dedicated regression testing module for convenience
