+++ halotools/empirical_models/composite_models/hod_models/tests/test_zheng07.py
@@ -116,20 +116,3 @@ def test_mass_definition_flexibility():
         halocat.halo_table["halo_mvir"]
     )
     model.populate_mock(halocat, seed=43)
-
-
-def test_mass_definition_flexibility2():
-    """Regression test for Issue #993."""
-    halocat = FakeSim()
-    halocat.halo_table["halo_mass_custom"] = np.copy(halocat.halo_table["halo_mvir"])
-    halocat.halo_table["halo_radius_custom"] = np.copy(halocat.halo_table["halo_rvir"])
-    halocat.halo_table.remove_column("halo_upid")
-    halocat.halo_table.remove_column("halo_mvir")
-    halocat.halo_table.remove_column("halo_rvir")
-    model = PrebuiltHodModelFactory(
-        "zheng07",
-        mdef="custom",
-        prim_haloprop_key="halo_mass_custom",
-        halo_boundary_key="halo_radius_custom",
-    )
-    model.populate_mock(halocat)
